Added gradient texture to the platforms

+++ game/character.h
@@ -48,7 +48,7 @@ public:
 
     int getHealth() const;
     void setHealth(int value);
-    void damaged(int power);
+    void damaged(qreal power);
 
     QVector<QString> State {
         "Attacking",
@@ -82,7 +82,7 @@ protected:
 
     int blockRadius = 25; // Default value, assigned dynamically in constructor
 
-    int damage = 0;
+    qreal damage = 0;
 
     short lookDirection = 1;
     bool isBlocking = false;
